fix: meta echo and safety

+++ library/Admin/Private/UserGroupSelector.php
@@ -2,11 +2,16 @@
 
 namespace Municipio\Admin\Private;
 
+use Municipio\HooksRegistrar\Hookable;
 use WpService\Contracts\AddAction;
 use WpService\Contracts\AddPostMeta;
+use WpService\Contracts\Checked;
+use WpService\Contracts\CurrentUserCan;
 use WpService\Contracts\DeletePostMeta;
 use WpService\Contracts\GetPostMeta;
 use WpService\Contracts\GetTerms;
+use WpService\Contracts\SanitizeTextField;
+use WpService\Contracts\WpVerifyNonce;
 
 /**
  * Represents a UserGroupSelector class.
@@ -14,7 +19,7 @@ use WpService\Contracts\GetTerms;
  * This class is responsible for handling user group selection functionality.
  * It is located in the file UserGroupSelector.php in the directory /workspaces/municipio-deployment/wp-content/themes/municipio/library/Admin/Private/.
  */
-class UserGroupSelector
+class UserGroupSelector implements Hookable
 {
     private string $userGroupMetaKey  = 'user-group-visibility';
     private string $userGroupTaxonomy = 'user_group';
@@ -22,7 +27,18 @@ class UserGroupSelector
     /**
      * Constructor for the UserGroupSelector class.
      */
-    public function __construct(private AddAction&DeletePostMeta&AddPostMeta&GetPostMeta&GetTerms $wpService)
+    public function __construct(private AddAction&DeletePostMeta&AddPostMeta&GetPostMeta&GetTerms&SanitizeTextField&WpVerifyNonce&CurrentUserCan&Checked $wpService)
+    {
+    }
+
+    /**
+     * Adds hooks for the UserGroupSelector class.
+     *
+     * This method adds hooks for the UserGroupSelector class.
+     *
+     * @return void
+     */
+    public function addHooks(): void
     {
         $this->wpService->addAction('post_submitbox_misc_actions', array($this, 'addUserVisibilitySelect'), 10);
         $this->wpService->addAction('attachment_submitbox_misc_actions', array($this, 'addUserVisibilitySelect'), 10);
@@ -43,6 +59,17 @@ class UserGroupSelector
      */
     public function saveUserVisibilitySelect($postId)
     {
+        if (
+            !isset($_POST['user_visibility_nonce']) ||
+            !$this->wpService->wpVerifyNonce($_POST['user_visibility_nonce'], 'save_user_visibility')
+        ) {
+            return;
+        }
+
+        if (!$this->wpService->currentUserCan('edit_post', $postId)) {
+            return;
+        }
+
         $this->wpService->deletePostMeta($postId, $this->userGroupMetaKey);
 
         if (empty($_POST[$this->userGroupMetaKey])) {
@@ -50,6 +77,7 @@ class UserGroupSelector
         }
 
         foreach ($_POST[$this->userGroupMetaKey] as $group) {
+            $group = $this->wpService->sanitizeTextField($group);
             $this->wpService->addPostMeta($postId, $this->userGroupMetaKey, $group, false);
         }
     }
@@ -70,7 +98,6 @@ class UserGroupSelector
         global $post;
 
         if (
-            empty($post->post_type) ||
             empty($terms = $this->wpService->getTerms(
                 [
                     'taxonomy'   => $this->userGroupTaxonomy,
@@ -82,24 +109,33 @@ class UserGroupSelector
         }
 
         $checked = $this->wpService->getPostMeta($post->ID, $this->userGroupMetaKey) ?: [];
-
-        echo '
-        <div id="' . $this->userGroupMetaKey . '" class="misc-pub-section" style="display: none;">
-            <label>' . __('User group visibility', 'municipio') . '</label>
-            <br><br>
-        ';
-
-        foreach ($terms as $term) {
-            echo '
-            <label style="display: block; margin-bottom: 5px;">
-                <input type="checkbox" name="' . $this->userGroupMetaKey . '[]" value="' . $term->slug . '" ' . (in_array($term->slug, $checked) ? 'checked' : '') . '>
-                ' . $term->name . '
-            </label>
-            ';
-        }
-
-        echo '
-        </div>
-        ';
+        $this->renderPrivateVisibilityList($terms, $checked);
     }
+
+    private function renderPrivateVisibilityList(array $terms, array $checked)
+{
+    echo sprintf(
+        '<div id="%s" class="misc-pub-section" style="display: none;">
+            <label>%s</label>
+            <br><br>',
+            $this->userGroupMetaKey,
+            __('User group visibility', 'municipio')
+    );
+
+    foreach ($terms as $term) {
+        echo sprintf(
+            '<label style="display: block; margin-bottom: 5px;">
+                <input type="checkbox" name="%s[]" value="%s" %s>
+                %s
+            </label>',
+            $this->userGroupMetaKey,
+            $term->slug,
+            $this->wpService->checked(in_array($term->slug, $checked), true, false),
+            $term->name
+        );
+    }
+
+    echo '</div>';
+}
+
 }
